fix(consonant-blocks): change 'background' to 'section-metadata'
